requirements.txt
Flask>=2.0,<4.0
sentence-transformers>=2.2.0,<3.0
scikit-learn>=1.0,<2.0
numpy>=1.20,<2.0
google-generativeai>=0.3.0,<0.6.0
# pickle, logging, os, re, json, time, math, enum, collections, typing, argparse are built-in
# Ensure your Python version is compatible with these libraries (e.g., Python 3.8+)